+++ dinput_hook/game_deltas/swrModel_delta.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "types.h"
+
+swrModel_Header *swrModel_LoadFromId_Hook(MODELID id);
